// backend/src/db/kysely-types.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type {ColumnType} from 'kysely';

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Numeric = ColumnType<number, number | string>;

export interface Ingredient {
  carbohydrates: Numeric;
  fat: Numeric;
  id: Generated<number>;
  name: string;
  protein: Numeric;
  servingSize: number;
  unit: string;
}

export interface Recipe {
  id: Generated<number>;
  name: string;
}

export interface RecipeIngredient {
  ingredientId: number;
  quantity: Numeric;
  recipeId: number;
  unit: string;
}

export interface RecipeInstruction {
  description: string;
  instructionId: Generated<number>;
  recipeId: number;
  stepNumber: number;
}

export interface RecipeTag {
  id: Generated<number>;
  recipeId: number;
  tag: string;
}

export interface DB {
  ingredient: Ingredient;
  recipe: Recipe;
  recipeIngredient: RecipeIngredient;
  recipeInstruction: RecipeInstruction;
  recipeTag: RecipeTag;
}